requirements.txt
absl-py==2.0.0
alembic==1.14.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
astunparse==1.6.3
Authlib==1.4.1
autopep8==2.3.1
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cryptography==44.0.0
cycler==0.11.0
distlib==0.3.9
Django==5.1.3
et_xmlfile==2.0.0
fastapi==0.109.1
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.42.0
fpdf==1.7.2
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.59.3
h5py==3.10.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Mako==1.3.8
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mercadopago==2.2.3
ml-dtypes==0.2.0
mpmath==1.3.0
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
networkx==3.2.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.3
pi==0.1.2
Pillow==10.0.0
platformdirs==4.3.6
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.11.4
secure-smtplib==0.1.1
six==1.16.0
sklearn==0.0.post11
sniffio==1.3.0
SQLAlchemy==2.0.37
sqlparse==0.5.2
starlette==0.35.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.27.1
Werkzeug==3.1.3
wrapt==1.14.1
